Expose nblib as library and add sample code
With this change, it is now possible to compile an external MD
program against nblib. A sample code computing forces and
performing integration with argon is added for demonstration
purposes. Since the sample code is linked to nblib, it is not
actually an external program, but it does compile as an
external program with the right linker and compiler flags.